Match DummyBehaviour to dummy locations
